[Adafruit_NeoPixel] Correct function signature issues causing compiler warnings

+++ lib/Adafruit_NeoPixel/esp8266.c
@@ -18,10 +18,10 @@ static inline uint32_t _getCycleCount(void) {
 
 #ifdef ESP8266
 IRAM_ATTR void espShow(
- uint8_t pin, uint8_t *pixels, uint32_t numBytes, __attribute__((unused)) boolean is800KHz) {
+ int16_t pin, uint8_t *pixels, uint32_t numBytes, __attribute__((unused)) boolean is800KHz) {
 #else
 void espShow(
- uint8_t pin, uint8_t *pixels, uint32_t numBytes, boolean is800KHz) {
+ int16_t pin, uint8_t *pixels, uint32_t numBytes, boolean is800KHz) {
 #endif
 
 #define CYCLES_800_T0H  (F_CPU / 2500001) // 0.4us
